Refactor for simplicity, and to support custom data types
PropMarshallers are now ObjectMarshallers, and handle raw objects rather than
props, which allows them to be used by other marshallers.

"MTProp" is massively simplified. There are 4 concrete prop classes:
the bool, int, float primitives, and MTObjectProp which covers everything else.
When an ObjectProp is being loaded, we look for a marshaller that can handle its
value type, and pass it off.

PageRef resolution needs to be reimplemented.