Merge all the XML stuff into the main library
So much of the loading process was tied up in the XML code, there
really wasn't any benefit to abstracting it out. I think the proper
way to handle this is probably to abstract at the data level rather
than the parsing level - e.g. an "MTElement" type that can have an
XML or a json implementation.