Move reader/writer creation to the marshaller.
This allows ListMarshaller to set true for the new isList parameter to addChild, a feature required
for the forthcoming JSON implementation. This was also needed (I think) to support lists of
primitives, but I haven't verified this... I just didn't see how it could be done based on the
ListMarshaller writeObject() code.